Bump version to 0.5.
Lots of cleanup of the way escaping is handled.

Refactored the Python code generator and the pretty printer to be more
consistent and cleaner.

Renamed the AST nodes in the Glop grammar to remove the Python-isms.

Added support for hex numbers, escaped hex literals, and
double-quoted strings.

Made it optional to separate rules with commas.

Refactored the code generation to not include rules and builtin functions
that aren't used in the grammar.

Cleaned up the APIs a bit.

Modified the code generator to also generate a main() function with
a simple input loop; this makes it much easier to test compiled
grammars and is closer to the final desired functionality for
the command line.