simplify the ASTs where possible, to simplify the compiler
The initial AST is unnecessarily verbose. For example, for

    grammar = "foo"

you end up with

    ['rules',
     ['rule',
      'grammar',
      ['choice',
       [['seq',
         [['lit',
           'foo']]]]]]]

when it could just be

    ['rules',
     ['rule',
      'grammar',
      ['lit',
       'foo']]]

by rewriting choices with one branch and sequences of one element.

By implementing these rules in the analyzer, this allows us to
make the compiler (and anything else that uses the AST) simpler.