add a basic benchmark
This adds a simple benchmark script that tests how long
it takes to parse the glop grammar itself, create an
interpreter for that grammar, and then use the interpreter
to re-parse the grammar.

The benchmark reports how many times glop can do this
in ten seconds; the actual duration of the benchmark
is configurable, but the result will be re-normalized
to 10s.