Fix error reporting on partial matches.
If you had a rule like `(anything:x ?( is_unicat('Ll')))*`, and the
rule failed, the failing position would be off by one, e.g.:
'a12' would report 'Unexpected "2" at column 3' instead of
'Unexpected '1' at column 2', because we would advance the err position
when the `anything` rule succeeded, and not rewind it back when the semantic
check failed.

Also, if you had a rule like `'abc'` and the rule failed on 'abd', we
would report the failure as occurring at the start of the string (column 1),
rather than at the character that actually failed (column 3).

This CL fixes both cases and adds tests for them.