Fix error reporting for nested rules.
This modifies the fix in f259ff, which wasn't quite
right. We were incorrectly rolling back the `errpos` field
when rules failed, which would lead to errors being
reported as having happened at the beginning of a rule,
rather than at the spot where something actually
happened.

In debugging this, I realized that part of the problem is
that error reporting for semantic predicate failures doesn't
quite work right, so I've filed a new bug for that and
updated the tests to be clearer.

See https://github.com/dpranke/pyjson5/issues/26 and
https://github.com/dpranke/glop/issues/3 for more on
this.