Remove char class support.
One of the changes I did on the dev branch was add support for
specifying char classes directly as part of a string, e.g.
'\d' would work like you'd expect from a regex engine (match any
digit).

However, the way I implemented this would've create a duplicate
sub-language if I'd done it "right". For example, '\d+' would
need to be the same as '\d'+, and if you had 'foo\d+bar' that
would be the same as 'foo' \d'+ 'bar'. But I didn't implement
that latter support for '+' inside the string literal. This
would be confusing to a user. And implementing it *inside*
the literal would duplicate the features *outside* the literal.
Also, in the compiler, it'd slurp in support for *all* the
char classes into a literal, you couldn't specify which ones
you wanted. The latter flaw seems fatal IMO.

This change removes the syntax for char classes inside string
literals for now. I will probably want to bring them back in
some form in the future but probably using a different syntax
and mechanism that plays more nicely with the other features
and can be compiled in in a minimal way.