Finish getting recursion working in the interpreter and add `number()`.
This change fixes the bug in the interpreter's implmentation of
`choice` that was keeping it from working correctly (which was then
leading to the left-, right- and both-recursion tests failing).

With this change, all tests pass for both the interpreter and the
compiler.

This change also adds the `number` builtin that takes a string and
returns either a python int or a python float (to match what the
python JSON library will do) as appropriate for numeric literals.
Adding `number` makes writing expression grammar tests a bit more
useful.