Start reworking compiler.py to be simpler, more reusable.
The current implementation of compile.py is hard to understand
and has a weird split between code (compile.py) and templates
(python_templates.py).

Ideally the code should be both simple and reusable between
(host_language, target_language) combinations, but it is
not clear what that would look like (whether a simple set
of templates can be host language independent).

This change checkpoints the work to refactor the compiler
code to be simple and straightforward. Once we have a
better sense of what <host_language, target_language>
combinations look like we can try to abstract things away
better.