Remove box formatting code as it is not needed.
The compiler code does not generate box instructions
with the new compiler rewrite, so the box code is all
dead and this CL removes it.

It might still be useful for something in the future
so for ease of finding it again I've created a
'box_formatting' branch that will still have it.