Add code to start inlining method bodies where possible.
Previously, a rule like `foo = bar "x"` would
compile to something like

    def _r_foo_0(self):
      self._h_seq([self._r_foo_1, self._r_foo_2])

and the definitions of ._r_foo_1 and ._r_foo_2
would be trivial, very short methods themselves.

This changes starts inlining those methods as either direct
invocations (in the case of `apply`) or lambda functions
(in the case of something like `lit`). This should speed
up and simplify the overall generated code.

You now end up with something like:

  def _r_foo_0(self):
    self._h_seq([
      self.bar_0,
      lambda: self._h_str('x)
    ])

This has not yet been implemented consistently for all AST node
types, but it modifies `choice`, `scope`, `seq`, and `not` to
prove the concept.