Checkpoint work on refactoring compiler.py.
Nearly all of the generated methods in a Parser class are of
the form

    def foo(self):
        bar

and yet we repeat that in every generating method (like _apply, etc).
We can refactor this by pushing the boilerplate up into
_gen_rule_text, and eventually up into handle_subrules.

This change, however, just converts _action and _apply to start
on the work. Subsequent changes will convert the other methods
as well.