Rework line-breaking for lists of arguments.
Prior to this change, line-breaking for lists of arguments
had some magic constants scattered around the file, and
didn't actually break lines as well as it could have.

Now we'll use the following logic:

-   If all the arguments fit on a single line with enough space for
    the function name and indentation (leaving 30 chars for that),
    put them all on one line:

        def foo(self, node):
            self.bar(arg1, arg2, arg3, ...)

-   If all the arguments fit on one line w/ proper indentation,
    use that next:

	def foo(self, node):
            self.bar(
                arg1, arg2, arg3, ...
            )

-   Failing those two, format the args one per line:
        def foo(self, node):
            self.bar(
                arg1,
                arg2,
                arg3,
                ...
            )

There is one more indentation practice that could be implemented,
for when you're between the second and third options, using multiple
arguments per line:

    def foo(self, node):
        self.bar(
            arg1, arg2,
            arg3, ...
        )

For now, this option isn't used but we can revisit it.