Fix logic in _star() to get correct coverage info.
In the generated _star() combinator, the logic used to read:

    while not self.failed:
         ...do stuff...
         if self.failed():
             break
    self._succeed(...)

It was not possible for self.failed to be True before the while loop,
and it was not possible for it to be True after the loop completed, because
of the the break statement, so there was an unreachable branch. If
you ran branch coverage over this, you'd get a complaint about that.

This CL changes the logic to just be `while True:` (which causes the
complaint to go away).